rename file and update test
